fixed a bug where the server would crash when a player died without an item (null pointer, duh). adding in the NEW game score GUI, which keeps track of the current game score in a moderately nice fashion on the client side, based off of messages from the server. started to implement stats tracking; only keeping track of CTF based stuff for now, will do kills and deaths soon! Going to start work soon on a new power up called RageSerum (damage booster)! Exciting shit.

+++ demo/src/main/java/com/cycapservers/game/Player.java
@@ -20,8 +20,10 @@ public class Player extends GameCharacter {
 	
 	public void die() {
 		this.isDead = true;
-		this.item_slot.drop();
-		this.item_slot = null;
+		if(this.item_slot !=  null) {
+			this.item_slot.drop();
+			this.item_slot = null;
+		}
 		this.x = -256;
 		this.y = -256;
 		this.lastDeathTime = System.currentTimeMillis();
